[subscriptions] fixes max interval for daily schedules

+++ database/migrations/2016_10_20_153402_create_schedules.php
@@ -30,7 +30,7 @@ class CreateSchedules extends Migration
 
             $table->string(Schedule::PERIOD, 15);
 
-            $table->tinyInteger(Schedule::INTERVAL)
+            $table->integer(Schedule::INTERVAL)
                   ->nullable();
 
             $table->integer(Schedule::ANCHOR)
